refactor: move GET_All_TICKETS query to a separate file
