+++ src/lib/apollo/query.tsx
@@ -1152,3 +1152,79 @@ export const VERIFY_TELEGRAM_QUEST = gql(/* GraphQL */ `
     }
   }
 `);
+
+export const GET_All_TICKETS = gql(`
+  query AllTickets(
+    $sort: TicketSortType
+    $status: TicketStatusType
+    $eventTypes: [EventType!]
+    $isVisibleOnly: Boolean
+    $limit: Int
+    $skip: Int
+  ) {
+    tickets(
+      sort: $sort
+      status: $status
+      eventTypes: $eventTypes
+      isVisibleOnly: $isVisibleOnly
+      limit: $limit
+      skip: $skip
+    ) {
+      _id
+      beginTime
+      untilTime
+      completed
+      description
+      description_v2 {
+        contentFormatType
+        contentEncodingType
+        content
+      }
+      participants {
+        _id
+        name
+        profileImageUrl
+      }
+      imageUrl
+      quests {
+        _id
+        title
+        title_v2 {
+          contentFormatType
+          contentEncodingType
+          content
+        }
+        description
+        questPolicy {
+          context
+          questPolicy
+        }
+      }
+      project {
+        _id
+        categories
+        description
+        imageUrl
+        name
+        projectSocial {
+          discordUrl
+          officialUrl
+          telegramUrl
+          twitterUrl
+          mediumUrl
+          naverBlogUrl
+        }
+      }
+      rewardPolicy {
+        context
+        rewardPoint
+        rewardPolicyType
+      }
+      title
+      winners {
+        name
+      }
+      visible
+    }
+  }
+`);
